Refactor comments to make it easier to add 'hello world'
The 'hello world' functionality will not be language-specific,
and it is not based on the code of the submission. It makes sense
to have the comments/ dir contain subdirectories based on what
functionality they refer to.

The functionality for the comments themselves (reading a text file
from a particular location on disk) is not related to the type of
feedback (analysis vs encouragement vs perhaps other things later).
By removing the knowledge of the language/detected issue from the
Comment struct itself, the comment can be used in a broader range of
circumstances.